# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/models/app/{app}" : {
      "delete" : {
        "tags" : [ "models" ],
        "summary" : "Delete existing models",
        "description" : "Delete existing models",
        "operationId" : "deleteModels",
        "parameters" : [ {
          "name" : "app",
          "in" : "path",
          "description" : "Application",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "ac", "ad", "rca" ],
            "example" : "ac"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful delete models"
          }
        }
      }
    },
    "/models/{modelId}" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get the details of a model",
        "description" : "Get the details of a model",
        "operationId" : "getModelStats",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "content" : {
                      "type" : "string",
                      "example" : ""
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["models"],
        "summary": "Update existing model ID",
        "description": "Renames the model with the given ID to a new ID",
        "operationId": "updateModelId",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "ID of the model to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New model ID",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newModelId": {
                    "type": "string",
                    "description": "New ID for the model"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Successful update name of a model"
          }
        }
      },
      "delete" : {
        "tags" : [ "models" ],
        "summary" : "Delete a model",
        "description" : "Delete a model",
        "operationId" : "deleteModel",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful delete a model"
          }
        }
      }
    },
    "/models/{modelId}/download" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Download a model file",
        "description" : "Download a model file",
        "operationId" : "downloadModel",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "multipart/form-data" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Model file does not exist"
          }
        }
      }
    },
    "/models/{modelId}/datasets/" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get all datasets of the model",
        "description" : "Get all datasets of the model",
        "operationId" : "getModelDatasets",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "datasets" : {
                      "type" : "array",
                      "items" : {
                        "type" : "string",
                        "example" : "rsl_poisoned_dataset.csv"
                      }
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Model does not exist"
          }
        }
      }
    },
    "/models/{modelId}/datasets/{datasetType}/download" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Download a dataset",
        "description" : "Download a dataset in .csv format",
        "operationId" : "downloadModelDatasets",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        }, {
          "name" : "datasetType",
          "in" : "path",
          "description" : "Type of the dataset",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "train", "test" ],
            "example" : "train"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "multipart/form-data" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Dataset file does not exist"
          }
        }
      }
    },
    "/models/{modelId}/confusion-matrix" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get the confusion matrix of the building model process",
        "description" : "Get the confusion matrix of the building model process",
        "operationId" : "getModelConfMatrix",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "content" : {
                      "type" : "string",
                      "example" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/models/{modelId}/build-config" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get the information relate to the building of the model",
        "description" : "Get the configuration of the model building process",
        "operationId" : "getModelBuildConfig",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/buildConfig"
                }
              }
            }
          }
        }
      }
    },
    "/models/{modelId}/predictions" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get predictions of the model",
        "description" : "Get predictions of the model",
        "operationId" : "getModelPredictions",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : ""
                }
              }
            }
          }
        }
      }
    },
    "/models/{modelId}/probabilities" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get probabilities of the model",
        "description" : "Get probabilities of the model",
        "operationId" : "getModelProbabilities",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : ""
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "trainingParameters" : {
        "type" : "object",
        "properties" : {
          "nb_epoch_cnn" : {
            "type" : "number",
            "example" : 5
          },
          "nb_epoch_sae" : {
            "type" : "number",
            "example" : 3
          },
          "batch_size_cnn" : {
            "type" : "number",
            "example" : 16
          },
          "batch_size_sae" : {
            "type" : "number",
            "example" : 32
          }
        }
      },
...
      "buildConfig" : {
        "type" : "object",
        "properties" : {
          "buildConfig" : {
            "type" : "object",
            "properties" : {
              "datasets" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "datasetId" : {
                      "type" : "string",
                      "example" : "report-bot-test-9bf49bb1-055f-45e6-bd05-dbfeb39f93bd"
                    },
                    "isAttack" : {
                      "type" : "boolean",
                      "example" : true
                    }
                  }
                }
              },
              "training_ratio" : {
                "type" : "number",
                "example" : 0.7
              },
              "training_parameters" : {
                "$ref" : "#/components/schemas/trainingParameters"
              }
            }
          }
        }
      },
...
    }
  }
}